# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Documentación automática del backend de Taskly",
        "title": "Taskly API",
        "contact": {},
        "version": "1.0.0"
    },
    "host": "localhost:7000",
    "basePath": "/",
    "paths": {
...
        "/equipos/{id}/miembros": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retorna la lista de usuarios que son miembros de un equipo específico, incluyendo su rol y datos básicos.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Miembros"
                ],
                "summary": "Obtener miembros de un equipo",
                "parameters": [
                    {
                        "type": "integer",
                        "example": 1,
                        "description": "ID del equipo",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de miembros obtenida exitosamente",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Miembro"
                            }
                        }
                    },
                    "400": {
                        "description": "ID de equipo inválido",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "No autenticado",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Equipo no encontrado",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Crea un registro de miembro y lo asocia a un equipo existente. Solo el propietario del equipo puede realizar esta acción.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Miembros"
                ],
                "summary": "Agregar un miembro a un equipo",
                "parameters": [
                    {
                        "type": "integer",
                        "example": 1,
                        "description": "ID del equipo",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Datos del miembro a agregar",
                        "name": "miembro",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Miembro"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Miembro agregado exitosamente",
                        "schema": {
                            "$ref": "#/definitions/models.Miembro"
                        }
                    },
                    "400": {
                        "description": "Datos inválidos o ID incorrecto",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "No autenticado",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Equipo no encontrado",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/equipos/{id}/miembros/{usuarioId}": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Remueve permanentemente la relación de un usuario con un equipo. Solo el propietario del equipo puede realizar esta acción.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Miembros"
                ],
                "summary": "Eliminar un miembro de un equipo",
                "parameters": [
                    {
                        "type": "integer",
                        "example": 1,
                        "description": "ID del equipo",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "example": 5,
                        "description": "ID del usuario a eliminar",
                        "name": "usuarioId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Miembro eliminado exitosamente",
                        "schema": {
                            "$ref": "#/definitions/models.MessageResponse"
                        }
                    },
                    "400": {
                        "description": "IDs inválidos o formato incorrecto",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "No autenticado",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Permisos insuficientes",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Miembro no encontrado en este equipo",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "gorm.DeletedAt": {
            "type": "object",
            "properties": {
                "time": {
                    "type": "string"
                },
                "valid": {
                    "description": "Valid is true if Time is not NULL",
                    "type": "boolean"
                }
            }
        },
        "models.Equipo": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "descripcion": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "miembros": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Miembro"
                    }
                },
                "nombre": {
                    "type": "string"
                },
                "ownerid": {
                    "type": "integer"
                },
                "tareas": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Tarea"
                    }
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "models.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 400
                },
                "error": {
                    "type": "string",
                    "example": "Datos inválidos"
                },
                "message": {
                    "type": "string",
                    "example": "Campo requerido"
                }
            }
        },
...
        "models.MessageResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Operación exitosa"
                }
            }
        },
        "models.Miembro": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "equipo": {
                    "$ref": "#/definitions/models.Equipo"
                },
                "equipoid": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "rol": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "usuario": {
                    "$ref": "#/definitions/models.Usuario"
                },
                "usuarioid": {
                    "type": "integer"
                }
            }
        },
...
        "models.Tarea": {
            "type": "object",
            "properties": {
                "asignadoid": {
                    "type": "integer"
                },
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "descripcion": {
                    "type": "string"
                },
                "equipoid": {
                    "type": "integer"
                },
                "estado": {
                    "type": "string"
                },
                "fechaLimite": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "prioridad": {
                    "type": "string"
                },
                "titulo": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
...
        "models.Usuario": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "email": {
                    "type": "string"
                },
                "equipos": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Equipo"
                    }
                },
                "id": {
                    "type": "integer"
                },
                "nombre": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        }
    }
}